Arrow_buttons_visible_new_component
