+++ frontend/src/components/RoleQuestionButtons.jsx
@@ -1,0 +1,66 @@
+import { Button, IconButton, Stack, Typography } from '@mui/material'
+import ArrowCircleRightIcon from '@mui/icons-material/ArrowCircleRight'
+import ArrowCircleLeftIcon from '@mui/icons-material/ArrowCircleLeft'
+import PropTypes from 'prop-types'
+
+export default function RoleQuestionButtons({
+    questionId,
+    totalQuestions,
+    isLastQuestion,
+    handleSubmit,
+}) {
+    return (
+        <Stack alignItems={'center'} paddingTop={4}>
+            <Stack
+                direction="row"
+                justifyContent="space-evenly"
+                alignItems="center"
+                spacing={4}
+                paddingY={1}
+            >
+                <IconButton
+                    data-testid="previous-question"
+                    aria-label="previous question"
+                    href={`/kysymys/${questionId - 1}`}
+                    disabled={questionId <= 1}
+                >
+                    <ArrowCircleLeftIcon fontSize="large" />
+                </IconButton>
+                {isLastQuestion ? (
+                    <Button
+                        variant="contained"
+                        color="secondary"
+                        onClick={handleSubmit}
+                    >
+                        Lopeta kysely
+                    </Button>
+                ) : (
+                    <Typography data-testid="current-progress">
+                        {questionId}/{totalQuestions}
+                    </Typography>
+                )}
+                <IconButton
+                    data-testid="next-question"
+                    aria-label="next question"
+                    href={`/kysymys/${questionId + 1}`}
+                    disabled={!totalQuestions || questionId >= totalQuestions}
+                >
+                    <ArrowCircleRightIcon fontSize="large" />
+                </IconButton>
+            </Stack>
+        </Stack>
+    )
+}
+
+const questionProps = PropTypes.shape({
+    id: PropTypes.number.isRequired,
+    content: PropTypes.string.isRequired,
+    options: PropTypes.array,
+})
+
+RoleQuestionButtons.propTypes = {
+    questionId: questionProps.number.isRequired,
+    totalQuestions: questionProps.number.isRequired,
+    isLastQuestion: questionProps.bool.isRequired,
+    handleSubmit: questionProps.func.isRequired,
+}
